dockerize/requirements.txt
langgraph>=0.2.40
langgraph-api>=0.2.3
langgraph-sdk>=0.1.61
langchain-openai>=0.2.0
langchain-core>=0.3.0
typing-extensions>=4.8.0
httpx
python-dotenv
openai
mcp
langchain-mcp-adapters
pandas
sqlalchemy
requests
psycopg2-binary